feat: add support for LTI launch refactor in xblock-lti-consumer
This commit updates the sample LTI proctoring integration to use the new LTI launch refactor in xblock-lti-consumer (https://github.com/openedx/xblock-lti-consumer/pull/288) and the Proctoring Services implementation (https://github.com/openedx/xblock-lti-consumer/pull/297).

As a reminder, this application currently serves as a mock implementation of the proctoring flow to demonstrate how we might use the xblock-lti-consumer library for proctoring. It is not intended to be deployed as-is or to serve as an actual proctoring flow implementation.

Here is a digest of important changes.

* The launch_proctoring, start_assessment, public_keyset, and access_token views are removed.
* The start_proctoring and end_assessment views are added.
* A signal handler for the LTI_1P3_PROCTORING_ASSESSMENT_STARTED is added.
* The lti AppConfig is modified to support the above signal.