fix: omit pattern names in .coveragerc file due to upgrade to 7.X.X
Version 7.0.0 of the coverage package includes a breaking change that changes the way that patterns are interpreted in the .coveragerc file. Below is a quote from the coverage CHANGELOG that describes the relevant change.

"Previously, * would incorrectly match directory separators, making precise matching difficult. This is now fixed, closing issue 1407."

This change causes the coverage on the project to drop -5.58%. This is because the omit file name pattern includes the following patterns in the .coveragerc file.

omit =
    edx_exams/settings*

settings is a directory, and prior to this change, the * was matching the directory separator. Because of this, files and subdirectories in this directory were correctly not included in the coverage calculation. After the upgrade to 7.X.X of the coverage package, * no longer matches the directory separator, so the files and subdirectories in this directory was incorrectly being included in the coverage calculation.

This commit corrects the omit file patterns to exclude the settings directory.