fix: add passed_due_date key to response from CourseExamAttemptView GET endpoint
The proctoring frontend hosted by the frontend-lib-special-exams frontend library is meant to block a learner from starting an exam that is passed due. The frontend relies on a key returned by the backend called passed_due_date to determine whether or not to block access to the exam. The frontend calls edx-exams's CourseExamAttemptView GET endpoint to get information about the exam and the attempt, and it expects this key to be present in the response.

However, this key was missing, which allowed learners to attempt to start the exam in the browser. Even though the backend correctly prevented them from starting the exam, the experience was confusing. This commit fixes this bug by sending a passed_due_date key in the response.