MST-1789 downstream triggers (#191)
* feat: add openedx_events and edx_event_bus_kafka to INSTALLED_APPS

* feat: install confluent_kafka[avro,schema-registry] for use with edx-event-bus-kafka

* feat: add event bus and Kafka Django settings

* feat: make User.full_name field non-nullable

This commit makes the User.full_name field non-nullable to ensure that there is only one representation of the empty value - the empty string. Currently, empty values for this field can be either None or the empty string. Because we're transitioning from a nullable to non-nullable state, a default is required to handle existing rows with a null value.

This is best practice; Django discourages setting both null=True and blank=True on CharField model fields. Furthermore, this was required by our event bus work. If an empty value is represented by None, this causes issues with the event bus, because None is not JSON serializable. Instead of converting a None value to the empty string in the event producer, correcting the the model definition is a better approach.

* feat: upgrade openedx-events to 8.8.0

* feat: emit the EXAM_ATTEMPT_SUBMITTED Open edX event when an exam is submitted

* feat: emit verified event for attempt (#186)

* feat: emit rejected exam signal (#181)

* feat: add errored event producer (#189)

* chore: fix upgrades

* chore: remove confluent-kafka import

* docs: local event bus development

* fix: remove local Kafka settings

This commit removes Django settings from the local settings file, used by the local application server, related to setting up the Kafka implementation of the event bus. This is because the event bus does not work outside of a Docker container. This is because the event bus is run through the devstack networking layer, which is inaccessible by the local application server.

* feat: update topic names for events (#193)

* feat: exam reset producer (#196)

send event to the event bus when an exam attempt is reset

---------

Co-authored-by: michaelroytman <mroytman@edx.org>
Co-authored-by: Michael Roytman <mroytman@2u.com>
Co-authored-by: Zachary Hancock <zhancock@edx.org>