feat: add escalation email to ProctoringSettingsView
This commit renames the CourseProviderSettingsView view to ProctoringSettingsView to accommodate additional settings beyond just provider settings; this more closely mirrors the purpose of the analogous edx-proctoring view. This commit also adds the escalation email to the response under the proctoring_escalation_email key. This changes allows the escalation email to appear in the error interstitial in the learning MFE.