fix: remove references to edx-drf-extensions USE-JWT-COOKIE constant
This commit removes references to the edx-drf-extensions USE-JWT-COOKIE constant. This constant was removed in version 10.0.0 of the edx-drf-extensions library. This constant is no longer necessary, because JWT authentication will now be attempted on all requests. This removes the need to set this constant explicitly inthe presence of a JWT cookie. This also removes the need to force an early return for lti endpoints, because edx-drf-extensions will fallback to other forms of authentication in the event that JWT authentication fails.

Please see the associated DEPR issue in the edx-drf-extensions library for more information: https://github.com/openedx/edx-drf-extensions/issues/371.