feat: unblock edx-drf-extensions upgrade (#267)
allows jwt/session user mismatch. This occurs when a user does an LTI launch as user A and than switches their account to user B while that session token is still active. In this case we should always use the JWT first but allow the mismatch during middleware. If another LTI launch occurs a new session is minted during that request.